validator/requirements.txt
# Core dependencies
duckdb>=0.9.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# API and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Data visualization and reporting
plotly>=5.17.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
kaleido>=0.2.1  # For static image export
reportlab>=4.0.0  # For PDF reports
openpyxl>=3.1.0   # For Excel reports
python-docx>=1.1.0  # For Word reports

# Configuration and utilities
pyyaml>=6.0.1
python-json-logger>=2.0.7
psutil>=5.9.0

# Async and concurrent processing
asyncio-mqtt>=0.16.0
aiofiles>=23.2.1

# Data validation and quality
pydantic-settings>=2.1.0
great-expectations>=0.18.0

# Optional: Supabase integration
supabase>=2.3.0

# Optional: tRPC integration
@trpc/server>=0.11.0
@trpc/client>=0.11.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0
